titan_v2/backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.115.0
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
# Data processing and analysis
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
# Environment and configuration
python-dotenv==1.0.0
pydantic==2.10.6
# Progress bars and utilities
tqdm==4.66.1
# File handling and path utilities
pathlib2==2.3.7
# AI and agent framework
camel-ai==0.2.74
# Jupyter dependencies for CodeExecutionToolkit
jupyter-client==8.6.0
jupyter-core==5.7.1
ipykernel==6.29.0